[bug-fix] scoreのupdate判定に使用していたUpdatedAtの関数が正しく動作していなかったのを修正

+++ crates/model/src/score/snapshot.rs
@@ -25,7 +25,7 @@ impl SnapShot {
         score: i32,
         combo: i32,
         minbp: i32,
-        timestamp: i32,
+        timestamp: i64,
     ) -> SnapShot {
         SnapShot {
             clear_type: ClearType::from_integer(clear_type),
